Video aspect ratio from renderer (#62)
* fixed aspect ratio with clipping on small screen

* using frame size from videoRenderer

* logs removed

* cleanup

* aspect ratio calculation unit tests added

* Dimensions.swift added

+++ RTSViewer/StreamingScreen/StreamingScreen.swift
@@ -28,7 +28,7 @@ struct StreamingScreen: View {
         BackgroundContainerView {
             ZStack {
                 let screenRect = UIScreen.main.bounds
-                let (videoFrameWidth, videoFrameHeight) = viewModel.calculateVideoViewWidthHeight(screenWidth: Double(screenRect.size.width), screenHeight: Double(screenRect.size.height))
+                let (videoFrameWidth, videoFrameHeight) = viewModel.calculateVideoViewWidthHeight(screenWidth: Float(screenRect.size.width), screenHeight: Float(screenRect.size.height))
                 VideoRendererView(uiView: viewModel.streamingView)
                     .frame(width: videoFrameWidth, height: videoFrameHeight)
